cleanup and type docs
