+++ lib/components/GPTDevIcon/GPTDev.tsx
