+++ lib/components/GPTDevIcon/GPTDevIcon.tsx
