+++ lib/components/GPTSlot/GPTSlot.type.ts
@@ -1,0 +1,5 @@
+import type { SlotUnit } from '../../types';
+
+export type GPTSlotProps = SlotUnit;
+
+export type GPTSlotInternalProps = GPTSlotProps & { isLoaded: boolean };
