+++ lib/index.ts
@@ -1,16 +1,17 @@
-import { GPTProvider } from './contexts/GPTProvider';
-import { GPTSlot } from './components/GPTSlot';
-import { useGPT } from './hooks/useGPT';
-import { GUIDELINES } from './constants/guidelines';
+import type { GPTSlotProps } from './components/GPTSlot/GPTSlot.type';
+import type { GPTProviderProps } from './components/GPTProvider/GPTProvider.type';
+import type { UseGPTProps } from './hooks/useGPT/useGPT.type';
 import type {
-  UseGPTProps,
-  GPTSlotProps,
-  GPTProviderProps,
   SlotLoadEvent,
   SlotViewableEvent,
   SlotRenderEndedEvent,
   Sizes,
 } from './types';
+
+import { GPTProvider } from './components/GPTProvider/GPTProvider';
+import { GPTSlot } from './components/GPTSlot/GPTSlot';
+import { useGPT } from './hooks/useGPT/useGPT';
+import { GUIDELINES } from './constants/guidelines';
 
 export type {
   UseGPTProps,
